Refactor document review page to use server and client components
